+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This file is used to track unreleased changes, for a complete history, check the
 ### ⏳ Deprecated
 
 * Berglas support is deprecated and will be removed in the next release
+* Random secret generation is deprecated and will be removed in the next release
 
 ### 🛑 Removed